project/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
appdata==2.1.2
asgiref==3.5.0
async-timeout==4.0.3
attrs==23.1.0
autopep8==1.6.0
backports.entry-points-selectable==1.1.0
blis==0.7.10
boto3==1.24.42
botocore==1.27.42
cachetools==4.2.4
catalogue==2.0.9
certifi==2021.5.30
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.2.0
click==8.0.1
colored==1.4.3
confection==0.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
cymem==2.0.7
distlib==0.3.2
dj-database-url==0.5.0
Django==4.0.3
django-cors-headers==3.8.0
django-countries==7.3.2
django-rest-knox==4.2.0
django-rest-swagger==2.2.0
django-storages==1.12.3
django-versatileimagefield==3.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
dnspython==2.1.0
drf-yasg==1.21.3
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.6.0/en_core_web_sm-3.6.0-py3-none-any.whl#sha256=83276fc78a70045627144786b52e1f2728ad5e29e5e43916ec37ea9c26a11212
filelock==3.0.12
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
frozenlist==1.4.0
future==0.18.2
google-api-core==1.34.0
google-auth==1.35.0
google-cloud-core==1.7.3
google-cloud-translate==2.0.1
googleapis-common-protos==1.60.0
greenlet==1.1.1
grpcio==1.58.0
grpcio-status==1.48.2
gunicorn==20.1.0
hrm==0.0.1
idna==2.10
inflection==0.5.1
itsdangerous==2.0.1
itypes==1.2.0
Jinja2==3.0.1
jmespath==1.0.1
joblib==1.3.2
langcodes==3.3.0
MarkupSafe==2.0.1
multidict==6.0.4
murmurhash==1.0.9
numpy==1.25.2
openai==0.28.0
openapi-codec==1.3.2
packaging==21.3
pathy==0.10.2
Pillow==9.2.0
pipenv==2021.5.29
platformdirs==2.2.0
preshed==3.0.8
protobuf==3.20.3
psycopg2==2.9.3
psycopg2-binary==2.9.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.8.0
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
PyJWT==1.7.1
pymongo==3.12.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.6
python-magic==0.4.27
python-magic-bin==0.4.14
pytz==2021.3
PyYAML==5.4.1
pyzk==0.9
regex==2023.8.8
requests==2.25.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3transfer==0.6.0
scikit-learn==1.3.0
scipy==1.11.2
semantic-version==2.9.0
setuptools-rust==1.1.2
simplejson==3.17.6
six==1.16.0
smart-open==6.3.0
spacy==3.6.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SQLAlchemy==1.4.23
sqlparse==0.4.2
srsly==2.4.7
thinc==8.1.12
threadpoolctl==3.2.0
to-requirements.txt==1.3.0
toml==0.10.2
tqdm==4.66.1
typer==0.9.0
typing_extensions==4.7.1
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.11
virtualenv==20.7.2
virtualenv-clone==0.5.6
wasabi==1.1.2
Werkzeug==2.0.1
yarl==1.9.2
